+++ R/MoEClust.R
@@ -39,8 +39,7 @@
 #' \strong{Maintainer}: Keefe Murphy - <\email{keefe.murphy@@ucd.ie}>
 #' @references K. Murphy and T. B. Murphy (2018). Gaussian Parsimonious Clustering Models with Covariates. \emph{To appear}. <\href{https://arxiv.org/abs/1711.05632v2}{arXiv:1711.05632v2}>.
 #' @examples
-#' \dontrun{
-#' data(ais)
+#' \donttest{data(ais)
 #' 
 #' # Fit two models
 #' res1  <- MoE_clust(ais[,3:7], G=2, gating=~BMI, expert=~sex,
@@ -51,7 +50,7 @@
 #' # Compare the best model from each set of results
 #' (comp <- MoE_compare(res1, res2, optimal.only=TRUE))
 #' 
-#' # Procude a plot for the optimal model                                                   
+#' # Produce a plot for the optimal model                                                   
 #' plot(comp$optimal, what="gpairs")
 #'
 #' data(CO2data)
@@ -87,7 +86,7 @@
   if(interactive()) {
     packageStartupMessage(paste("\n___  ___      _____ _____ _           _   \n|  \\/  |     |  ___/  __ \\ |         | |     Gaussian Parsimonious \n| .  . | ___ | |__ | /  \\/ |_   _ ___| |_   Clustering Models with\n| |\\/| |/ _ \\|  __|| |   | | | | / __| __|        Covariates and a\n| |  | | (_) | |___| \\__/\\ | |_| \\__ \\ |_          Noise Component\n\\_|  |_/\\___/\\____/ \\____/_|\\__,_|___/\\__|           version", version, "\n"))
   } else   {
-    packageStartupMessage("\nPackage 'MoEClust' version ", version, ".")
+    packageStartupMessage("\nPackage ", sQuote("MoEClust"), " version ", version, ".\n")
   }
-    packageStartupMessage(paste("Type '?MoEClust' to see a brief guide to how to use this R package.\nType", sQuote(paste0("citation(", dQuote("MoEClust"),")")) ,"for citing the package in publications.\nType 'MoE_news()' to see new features recent changes and bug fixes.\n"))
+    packageStartupMessage(paste("Type", sQuote("?MoEClust"), "to see a brief guide to how to use this R package.\nType", sQuote(paste0("citation(", dQuote("MoEClust"),")")) ,"for citing the package in publications.\nType", sQuote("MoE_news()"), "to see new features recent changes and bug fixes.\n"))
 }